Redirected index to home page.
Updated Button on sign in page to "go back" for better clarification

Added sign in and sign up handler functions to redirect user to sign in or sign up.
Additionally, added conditional rendering to output sign in and sign up buttons or sign out and delete button for users with permanent accounts.

Added functionality to allow users to sign in as "anonynous" guests.